tmp: part of the communication protocol server-side

+++ highway-pursuit-gym/envs/_remote/highway_pursuit_data.py
@@ -19,7 +19,7 @@ class Instruction(ctypes.Structure):
 
 class Info(ctypes.Structure):
     _fields_ = (
-        ('tps', ctypes.c_uint),
+        ('tps', ctypes.c_float),
         ('memory', ctypes.c_float),
     )
